Fixup the finds to properly exclude vendor et al
I found that the previous `find` commands were not recursing into `vendor`, but were returning a `./vendor` entry even though it was not a `-type f`.  I found the article below, which gave another alternative and it seems to work where the current command falls short.

The addition of `./vendor` to the list was causing `misspell` to flag spelling errors within files in `vendor/` even though they were not fully enumerated.

See: https://linuxconfig.org/how-to-explicitly-exclude-directory-from-find-command-s-search